+++ discord/ext/test/__init__.py
@@ -3,7 +3,8 @@ __title__ = "dpytest"
 __author__ = "Rune Tynan"
 __license__ = "MIT"
 __copyright__ = "Copyright 2018-2019 CraftSpider"
-__version__ = "0.2.0"
+__version__ = "0.5.0"
 
 from . import backend
 from .runner import *
+from .verify import verify, Verify, VerifyMessage, VerifyActivity