+++ discord/ext/test/runner.py
@@ -1,6 +1,12 @@
 """
     Main module for setting up and running tests using dpytest.
-    Handles configuration of a bot, and setup of the discord environment
+    Handles configuration of a bot, and setup of the discord environment.
+
+    All public functions in this module are re-exported at ``discord.ext.test``, this is the primary
+    entry point for users of the library and most of what they should interact with
+
+    See also:
+        :mod:`discord.ext.test.verify`
 """
 
 
@@ -8,12 +14,11 @@ import sys
 import asyncio
 import logging
 import discord
-import discord.ext.commands as commands
 import typing
 import pathlib
 
 from . import backend as back, callbacks, _types
-from .utils import embed_eq
+from .utils import PeekableQueue
 
 
 class RunnerConfig(typing.NamedTuple):
@@ -30,23 +35,33 @@ class RunnerConfig(typing.NamedTuple):
 
 log = logging.getLogger("discord.ext.tests")
 _cur_config: typing.Optional[RunnerConfig] = None
-sent_queue: asyncio.Queue = asyncio.queues.Queue()
-error_queue: asyncio.Queue = asyncio.queues.Queue()
+sent_queue: PeekableQueue = PeekableQueue()
+error_queue: PeekableQueue = PeekableQueue()
 
 
 def require_config(func: typing.Callable[..., _types.T]) -> typing.Callable[..., _types.T]:
+    """
+        Decorator to enforce that configuration is completed before the decorated function is
+        called.
+
+    :param func: Function to decorate
+    :return: Function with added check for configuration being setup
+    """
     def wrapper(*args, **kwargs):
         if _cur_config is None:
             log.error("Attempted to make call before runner configured")
             raise RuntimeError(f"Configure runner before calling {func.__name__}")
         return func(*args, **kwargs)
     wrapper.__wrapped__ = func
+    wrapper.__annotations__ = func.__annotations__
+    wrapper.__doc__ = func.__doc__
     return wrapper
 
 
 async def run_all_events() -> None:
     """
-        Ensure that all dpy related coroutines have completed or been cancelled
+        Ensure that all dpy related coroutines have completed or been cancelled. If any dpy coroutines
+        are currently running, this will also wait for those.
     """
     while True:
         if sys.version_info[1] >= 7:
@@ -62,7 +77,8 @@ async def run_all_events() -> None:
 
 async def finish_on_command_error() -> None:
     """
-        Ensure that all dpy related coroutines have completed or been cancelled
+        Ensure that all dpy related coroutines have completed or been cancelled. This will only
+        wait for dpy related coroutines, not any other coroutines currently running.
     """
     if sys.version_info[1] >= 7:
         pending = filter(lambda x: x._coro.__name__ == "_run_event", asyncio.all_tasks())
@@ -73,98 +89,18 @@ async def finish_on_command_error() -> None:
             await task
 
 
-def verify_message(text: str = None, equals: bool = True, contains: bool = False, peek: bool = False, assert_nothing: bool = False) -> None:
-    """
-        Assert that a message was sent with the given text, or that a message was sent that *doesn't* match the
-        given text
-
-    :param text: Text to match, or None to match anything
-    :param equals: Whether to negate the check
-    :param assert_nothing: Whether to assert that nothing was sent at all
-    """
-    if text is None:
-        equals = not equals
-    if assert_nothing:
-        assert sent_queue.qsize() == 0, f"A message was not meant to be sent but this message was sent {sent_queue.get_nowait().content}"  # noqa: E501
-        return
-
-    try:
-        message = sent_queue.get_nowait()
-        if peek:
-            messages = [message]
-            while not sent_queue.empty():
-                tmp_message = sent_queue.get_nowait()
-                messages.append(tmp_message)
-            for tmp_message in messages:
-                sent_queue.put_nowait(tmp_message)
-        if equals:
-            if contains:
-                assert text in message.content, f"Didn't find expected text. Expected {text} to be in {message.content}"
-            else:
-                assert message.content == text, f"Didn't find expected text. Expected {text}, found {message.content}"
-        else:
-            if contains:
-                assert text not in message.content, f"Found unexpected text. Expected something not containing {text}"
-            else:
-                assert message.content != text, f"Found unexpected text. Expected something not matching {text}"
-    except asyncio.QueueEmpty:
-        raise AssertionError("No message returned by command")
-
-
 def get_message(peek: bool = False) -> discord.Message:
     """
-        Allow the user to retrieve a message sent by the bot
+        Allow the user to retrieve the most recent message sent by the bot
 
-    :param peek: do not remove the message from the queue of messages
-    :return:
+    :param peek: If true, message will not be removed from the queue
+    :return: Most recent message from the queue
     """
-    message = sent_queue.get_nowait()
     if peek:
-        messages = [message]
-        while not sent_queue.empty():
-            tmp_message = sent_queue.get_nowait()
-            messages.append(tmp_message)
-        for tmp_message in messages:
-            sent_queue.put_nowait(tmp_message)
-    return message
-
-
-def verify_embed(embed: discord.Embed = None, allow_text: bool = False, equals: bool = True, peek: bool = False, assert_nothing: bool = False) -> None:
-    """
-        Assert that a message was sent containing an embed, or that a message was sent not
-        containing an embed
-
-    :param embed: Embed to compare against
-    :param allow_text: Whether non-embed text is allowed
-    :param equals: Whether to invert the assertion
-    :param assert_nothing: Whether to assert that no message was sent
-    """
-    if embed is None:
-        equals = not equals
-    if assert_nothing:
-        assert sent_queue.qsize() == 0, f"A message was not meant to be sent but this message was sent {sent_queue.get_nowait().content}"  # noqa: E501
-        return
-
-    try:
+        message = sent_queue.peek()
+    else:
         message = sent_queue.get_nowait()
-        if not allow_text:
-            assert not message.content
-        if peek:
-            messages = [message]
-            while not sent_queue.empty():
-                tmp_message = sent_queue.get_nowait()
-                messages.append(tmp_message)
-            for tmp_message in messages:
-                sent_queue.put_nowait(tmp_message)
-        emb = None
-        if len(message.embeds) > 0:
-            emb = message.embeds[0]
-        if equals:
-            assert embed_eq(emb, embed), "Didn't find expected embed"
-        else:
-            assert not embed_eq(emb, embed), "Found unexpected embed"
-    except asyncio.QueueEmpty:
-        raise AssertionError("No message returned by command")
+    return message
 
 
 def get_embed(peek: bool = False) -> discord.Embed:
@@ -172,63 +108,21 @@ def get_embed(peek: bool = False) -> discord.Embed:
         Allow the user to retrieve an embed in a message sent by the bot
 
     :param peek: do not remove the message from the queue of messages
-    :return:
+    :return: Embed of the most recent message in the queue
     """
-    message = sent_queue.get_nowait()
+
     if peek:
-        messages = [message]
-        while not sent_queue.empty():
-            tmp_message = sent_queue.get_nowait()
-            messages.insert(0, tmp_message)
-        for tmp_message in messages:
-            sent_queue.put_nowait(tmp_message)
+        message = sent_queue.peek()
+    else:
+        message = sent_queue.get_nowait()
     return message.embeds[0]
 
 
-async def verify_file(file: typing.Union[str, pathlib.Path] = None, allow_text: bool = False, equals: bool = True, assert_nothing: bool = False) -> None:
-    if file is not None:
-        with file.open('rb') as f:
-            expected = f.read()
-    else:
-        equals = not equals
-        expected = None
-    if assert_nothing:
-        assert sent_queue.qsize() == 0, f"A message was not meant to be sent but this message was sent {sent_queue.get_nowait().content}"  # noqa: E501
-
-    try:
-        message = sent_queue.get_nowait()
-        if not allow_text:
-            assert not message.content
-
-        attach = None
-        if len(message.attachments) > 0:
-            attach = await message.attachments[0].read()
-        if equals:
-            assert attach == expected, "Didn't find expected file"
-        else:
-            assert attach != expected, "Found unexpected file"
-    except asyncio.QueueEmpty:
-        raise AssertionError("No message returned by command")
-
-
-def verify_activity(activity: discord.Activity = None, equals: bool = True) -> None:
-    if activity is None:
-        equals = not equals
-    me = _cur_config.guilds[0].me
-
-    me_act = me.activity
-    if isinstance(activity, discord.BaseActivity):
-        activity = (activity.name, activity.url, activity.type)
-    if isinstance(me_act, discord.BaseActivity):
-        me_act = (me_act.name, me_act.url, me_act.type)
-
-    if equals:
-        assert me_act == activity, "Didn't find expected activity"
-    else:
-        assert me_act != activity, "Found unexpected activity"
-
-
 async def empty_queue() -> None:
+    """
+        Empty the current message queue. Waits for all events to complete to ensure queue
+        is not immediately added to after running.
+    """
     await run_all_events()
     while not sent_queue.empty():
         await sent_queue.get()
@@ -236,37 +130,14 @@ async def empty_queue() -> None:
         await error_queue.get()
 
 
-async def message_callback(message: discord.Message) -> None:
+async def _message_callback(message: discord.Message) -> None:
+    """
+        Internal callback, on a message being sent (in any channel) adds it to the queue
+
+    :param message: Message sent on discord
+    """
     await sent_queue.put(message)
 
-
-async def error_callback(ctx: commands.Context, error: commands.CommandError) -> None:
-    await error_queue.put((ctx, error))
-
-
-async def create_role_callback(guild: discord.Guild, role: discord.Role, reason: str = None) -> None:
-    roles = [role] + guild.roles
-    if role.position == -1:
-        for r in roles:
-            if r.position != 0:
-                r.position += 1
-        role.position = 1
-    back.update_guild(guild, roles=roles)
-
-
-async def move_role_callback(guild: discord.Guild, role: discord.Role, positions: typing.List[_types.JsonDict], reason: str = None) -> None:
-    for pair in positions:
-        guild._roles[pair["id"]].position = pair["position"]
-
-
-async def add_role_callback(member: discord.Member, role: discord.Role, reason: str = None) -> None:
-    roles = [role] + [x for x in member.roles if x.id != member.guild.id]
-    back.update_member(member, roles=roles)
-
-
-async def remove_role_callback(member: discord.Member, role: discord.Role, reason: str = None) -> None:
-    roles = [x for x in member.roles if x != role and x.id != member.guild.id]
-    back.update_member(member, roles=roles)
 
 from itertools import count
 counter = count(0)
@@ -279,6 +150,15 @@ async def message(
         member: typing.Union[discord.Member, int] = 0,
         attachments: typing.List[typing.Union[pathlib.Path, str]] = None
 ) -> discord.Message:
+    """
+        Fake a message being sent by some user to a channel.
+
+    :param content: Content of the message
+    :param channel: Channel to send to, or index into the config list
+    :param member: Member sending the message, or index into the config list
+    :param attachments: Message attachments to include, as file paths.
+    :return: New message that was sent
+    """
     if isinstance(channel, int):
         channel = _cur_config.channels[channel]
     if isinstance(member, int):
@@ -319,9 +199,17 @@ async def set_permission_overrides(
         overrides: typing.Optional[discord.PermissionOverwrite] = None,
         **kwargs: typing.Any,
 ) -> None:
+    """
+        Set the permission override for a channel, as if set by another user.
+
+    :param target: User or Role the permissions override is being set for
+    :param channel: Channel the permissions are being set on
+    :param overrides: The permissions to use, as an object. Conflicts with using ``kwargs``
+    :param kwargs: The permissions to use, as a set of keys and values. Conflicts with using ``overrides``
+    """
     if kwargs:
         if overrides:
-            raise ValueError("either overrides parameter or kwargs")
+            raise ValueError("Must supply either overrides parameter or kwargs, not both")
         else:
             overrides = discord.PermissionOverwrite(**kwargs)
 
@@ -341,6 +229,12 @@ async def set_permission_overrides(
 
 @require_config
 async def add_role(member: discord.Member, role: discord.Role) -> None:
+    """
+        Add a role to a member, as if added by another user.
+
+    :param member: Member to add the role to
+    :param role: Role to be added
+    """
     if isinstance(member, int):
         member = _cur_config.members[member]
     if not isinstance(role, discord.Role):
@@ -352,6 +246,12 @@ async def add_role(member: discord.Member, role: discord.Role) -> None:
 
 @require_config
 async def remove_role(member: discord.Member, role: discord.Role) -> None:
+    """
+        Remove a role from a member, as if removed by another user.
+
+    :param member: Member to remove the role from
+    :param role: Role to remove
+    """
     if isinstance(member, int):
         member = _cur_config.members[member]
     if not isinstance(role, discord.Role):
@@ -361,20 +261,48 @@ async def remove_role(member: discord.Member, role: discord.Role) -> None:
     back.update_member(member, roles=roles)
 
 
-# self is a discord.Message based on how this framework runs
-async def simulate_reaction(self: discord.Message, emoji: str, member: discord.Member):
+@require_config
+async def add_reaction(user: typing.Union[discord.user.BaseUser, discord.abc.User], message: discord.Message, emoji: str) -> None:
+    """
+        Add a reaction to a message, as if added by another user
 
-    state = back.get_state()
-    await state.http.add_reaction(self.channel.id, self.id, emoji)
+    :param user: User who reacted
+    :param message: Message they reacted to
+    :param emoji: Emoji that was used
+    """
+    back.add_reaction(message, user, emoji)
     await run_all_events()
-
-    if not error_queue.empty():
-        err = await error_queue.get()
-        raise err[1]
 
 
 @require_config
-async def member_join(guild: int = 0, user: discord.User = None, *, name: str = None, discrim: typing.Union[str, int] = None) -> discord.Member:
+async def remove_reaction(user: typing.Union[discord.user.BaseUser, discord.abc.User], message: discord.Message, emoji: str) -> None:
+    """
+        Remove a reaction from a message, as if done by another user
+
+    :param user: User who removed their react
+    :param message: Message they removed react from
+    :param emoji: Emoji that was removed
+    """
+    back.remove_reaction(message, user, emoji)
+    await run_all_events()
+
+
+@require_config
+async def member_join(
+        guild: typing.Union[discord.Guild, int] = 0,
+        user: typing.Optional[discord.User] = None,
+        *,
+        name: str = None,
+        discrim: typing.Union[str, int] = None
+) -> discord.Member:
+    """
+        Have a new member join a guild, either an existing or new user for the framework
+
+    :param guild: Guild member is joining
+    :param user: User to join, or None to create a new user
+    :param name: If creating a new user, the name of the user. None to auto-generate
+    :param discrim: If creating a new user, the discrim of the user. None to auto-generate
+    """
     import random
     if isinstance(guild, int):
         guild = _cur_config.guilds[guild]
@@ -385,15 +313,31 @@ async def member_join(guild: int = 0, user: discord.User = None, *, name: str = 
         if discrim is None:
             discrim = random.randint(1, 9999)
         user = back.make_user(name, discrim)
+    elif name is not None or discrim is not None:
+        raise ValueError("Cannot supply user at the same time as name/discrim")
     member = back.make_member(user, guild)
     return member
 
 
 def get_config() -> RunnerConfig:
+    """
+        Get the current runner configuration
+
+    :return: Current runner config
+    """
     return _cur_config
 
 
 def configure(client: discord.Client, num_guilds: int = 1, num_channels: int = 1, num_members: int = 1) -> None:
+    """
+        Set up the runner configuration. This should be done before any tests are run.
+
+    :param client: Client to configure with. Should be the bot/client that is going to be tested.
+    :param num_guilds: Number of guilds to start the configuration with. Default is 1
+    :param num_channels: Number of text channels in each guild to start with. Default is 1
+    :param num_members: Number of members in each guild (other than the client) to start with. Default is 1.
+    """
+
     global _cur_config
 
     if not isinstance(client, discord.Client):
@@ -419,11 +363,7 @@ def configure(client: discord.Client, num_guilds: int = 1, num_channels: int = 1
     client.on_command_error = on_command_error
 
     # Configure global callbacks
-    callbacks.set_callback(message_callback, "send_message")
-    callbacks.set_callback(create_role_callback, "create_role")
-    callbacks.set_callback(move_role_callback, "move_role")
-    callbacks.set_callback(add_role_callback, "add_role")
-    callbacks.set_callback(remove_role_callback, "remove_role")
+    callbacks.set_callback(_message_callback, "send_message")
 
     back.get_state().stop_dispatch()