+++ tests/test_send.py
@@ -9,7 +9,7 @@ async def test_message(bot):
     channel = guild.text_channels[0]
 
     await channel.send("Test Message")
-    await dpytest.empty_queue()
+
 
 @pytest.mark.asyncio
 async def test_embed(bot):
@@ -20,4 +20,3 @@ async def test_embed(bot):
     embed.add_field(name="Field 1", value="Lorem ipsum")
 
     await channel.send(embed=embed)
-    await dpytest.empty_queue()